# docs/04-api-spec.yaml
openapi: 3.1.0
info:
  title: FlagOps API
  description: Feature Flag & Application Configuration Service
  version: 0.1.0
paths:
...
  /eval/v1/ruleset:
    get:
      tags:
      - Evaluation
      summary: Get Ruleset
      description: Download environment ruleset for SDK evaluation with ETag / 304
        Not Modified.
      operationId: get_ruleset_eval_v1_ruleset_get
      parameters:
      - name: environment_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Environment Id
      - name: key
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
      - name: If-None-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-None-Match
      - name: X-FlagOps-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Flagops-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /eval/v1/stream:
    get:
      tags:
      - Evaluation
      summary: Stream ruleset updates via SSE
      description: Server-Sent Events stream notifying SDKs when the ruleset version
        changes.
      operationId: stream_ruleset_eval_v1_stream_get
      parameters:
      - name: key
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
      - name: X-FlagOps-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Flagops-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /eval/v1/flags/{flag_key}/evaluate:
    post:
      tags:
      - Evaluation
      summary: Evaluate Single Flag
      description: Evaluate a single feature flag.
      operationId: evaluate_single_flag_eval_v1_flags__flag_key__evaluate_post
      parameters:
      - name: flag_key
        in: path
        required: true
        schema:
          type: string
          title: Flag Key
      - name: key
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
      - name: X-FlagOps-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Flagops-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /eval/v1/flags/evaluate-all:
    post:
      tags:
      - Evaluation
      summary: Evaluate All
      description: Evaluate all feature flags in the environment.
      operationId: evaluate_all_eval_v1_flags_evaluate_all_post
      parameters:
      - name: key
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
      - name: X-FlagOps-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Flagops-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluateAllResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /eval/v1/events:
    post:
      tags:
      - Evaluation
      summary: Ingest Events
      description: Ingest evaluation events batch with hashed context keys.
      operationId: ingest_events_eval_v1_events_post
      parameters:
      - name: key
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
      - name: X-FlagOps-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Flagops-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventBatchRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventBatchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /eval/v1/config/{namespace}:
    get:
      tags:
      - Evaluation
      summary: Fetch published configurations for namespace
      description: Fetch published configurations for SDK/Client with ETag / 304 Not
        Modified.
      operationId: get_config_eval_v1_config__namespace__get
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          title: Namespace
      - name: environment_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Environment Id
      - name: key
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
      - name: If-None-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-None-Match
      - name: X-FlagOps-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Flagops-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EvaluateAllResponse:
      properties:
        flags:
          additionalProperties:
            $ref: '#/components/schemas/EvaluationResponse'
          type: object
          title: Flags
      type: object
      required:
      - flags
      title: EvaluateAllResponse
      description: Result of evaluating all feature flags in an environment.
    EvaluationRequest:
      properties:
        context:
          additionalProperties: true
          type: object
          title: Context
          description: Evaluation context containing targetingKey/userId and user
            attributes
      type: object
      title: EvaluationRequest
      description: Payload for evaluating flags.
    EvaluationResponse:
      properties:
        flagKey:
          type: string
          title: Flagkey
        value:
          title: Value
        variant:
          anyOf:
          - type: string
          - type: 'null'
          title: Variant
        reason:
          type: string
          title: Reason
        errorCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Errorcode
        flagMetadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Flagmetadata
      type: object
      required:
      - flagKey
      - value
      - reason
      title: EvaluationResponse
      description: Result of evaluating a single feature flag.
    EventBatchRequest:
      properties:
        events:
          items:
            $ref: '#/components/schemas/EventItem'
          type: array
          minItems: 1
          title: Events
      type: object
      required:
      - events
      title: EventBatchRequest
      description: Batch of evaluation events to be ingested.
    EventBatchResponse:
      properties:
        status:
          type: string
          title: Status
          default: accepted
        count:
          type: integer
          title: Count
      type: object
      required:
      - count
      title: EventBatchResponse
      description: Response after ingesting event batch.
    EventItem:
      properties:
        flag_id:
          type: string
          format: uuid
          title: Flag Id
        variation_id:
          type: string
          format: uuid
          title: Variation Id
        reason:
          type: string
          maxLength: 30
          title: Reason
        context_key:
          type: string
          maxLength: 200
          minLength: 1
          title: Context Key
        context:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Context
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - flag_id
      - variation_id
      - reason
      - context_key
      title: EventItem
      description: Single evaluation event payload.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer